% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{addContrasts}
\alias{addContrasts}
\title{Add a set of new Contrasts to an existing Design}
\usage{
addContrasts(conn, designID, contrastMatrix, names = NULL,
  descriptions = NULL)
}
\arguments{
\item{conn}{Database connection}

\item{designID}{The integer ID of an existing Design}

\item{contrastMatrix}{The contrast matrix, likely returned by \code{\link[limma]{makeContrasts}}}

\item{names}{Names of the contrasts. If \code{NULL}, column names are used.}

\item{descriptions}{Descriptions of the contrasts. Must be the same length
as the column number of the contrast matrix. It can contain plain-text
descriptions of the contrasts.}
}
\value{
New contrast IDs are returned. Side-effect: the new contrasts are
 inserted into the \code{Contrasts} table.
}
\description{
Add a set of new Contrasts to an existing Design
}
\seealso{
\code{\link[limma]{makeContrasts}}
}

