% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{addDesign}
\alias{addDesign}
\title{Add a new design to an existing Dataset}
\usage{
addDesign(conn, datasetID, name = "New design", description = NA,
  sampleSubset, featureSubset, designMatrix)
}
\arguments{
\item{conn}{Database connection}

\item{datasetID}{The integer ID of an existing dataset}

\item{name}{Name of the design}

\item{description}{Description of the design}

\item{sampleSubset}{A \code{sampleSubset} data.frame}

\item{featureSubset}{A \code{featureSubset} data.frame}

\item{designMatrix}{Design matrix, probably returned from \code{model.matrix}}
}
\value{
The design ID is returned. Side-effect: the new design is inserted in
to the \code{Designs} table.
}
\description{
Add a new design to an existing Dataset
}

