% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{assertFeatureSubset}
\alias{assertFeatureSubset}
\title{Assert the correct structure of featureSubset}
\usage{
assertFeatureSubset(featureSubset)
}
\arguments{
\item{featureSubset}{A \code{data.frame}, see details.

A valid \code{featureSubset} is a \code{data.frame} object containing
three columns: \code{FeatureIndex} (integer), \code{isUsed} (logical),
and \code{Description} (character).
Note that the column names are case-sensitive}
}
\value{
Invisible \code{TRUE} if the object is valid, otherwise an error
is raised.
}
\description{
Assert the correct structure of featureSubset
}

