% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{assertSampleSubset}
\alias{assertSampleSubset}
\title{Assert the correct structure of sampleSubset}
\usage{
assertSampleSubset(sampleSubset)
}
\arguments{
\item{sampleSubset}{A \code{data.frame}, see details.

A valid \code{sampleSubset} is a \code{data.frame} object containing
three columns: \code{SampleIndex} (integer), \code{isUsed} (logical),
and \code{Description} (character).
Note that the column names are case-sensitive}
}
\value{
Invisible \code{TRUE} if the object is valid, otherwise an error
is raised.
}
\description{
Assert the correct structure of sampleSubset
}

