% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{blobs}
\alias{blobs}
\title{Serialize an object and make a blob out of it}
\usage{
blobs(object)
}
\arguments{
\item{object}{Any R object}
}
\value{
A blob object that can be unserialized when read from the database
}
\description{
Serialize an object and make a blob out of it
}
\examples{
myList1 <- list(list(myTeam="HSV", score=16), list(yourTeam="FCB", score=12))
myList2 <- list(list(myTeam="HSV", score=19), list(yourTeam="FCB", score=13))
myBlob1 <- blobs(myList1)
myBlob2 <- blobs(myList2)
myDf <- rbind(data.frame(Day=1, Object=myBlob1),
              data.frame(Day=2, Object=myBlob2))
myCon <- dbConnect(SQLite(), ":memory:")
dbWriteTable(conn=myCon, name="testBlobs", value=myDf)
myDfOut <- dbReadTable(conn=myCon, name="testBlobs")
myList1Out <- unserialize(myDfOut$Object[[1]])
myList2Out <- unserialize(myDfOut$Object[[2]])
stopifnot(identical(myList1, myList1Out))
stopifnot(identical(myList2, myList2Out))
}
