% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factors.R
\name{decodeFactor}
\alias{decodeFactor}
\title{Decode factors from plain character strings}
\usage{
decodeFactor(str)
}
\arguments{
\item{str}{A string vector

\code{decodeFactor} decodes a factor vector from a string vector that was
encoded by \code{encodeFactor}. This is especially useful when factors need
to be written into relational database management systems (RDMS) while it is
desired that the order of the levels is kept.}
}
\description{
Decode factors from plain character strings
}
\examples{
testFac <- factor(c("HBV", "FCB", "BVB", "HBV"), levels=c("HBV", "FCB", "BVB"))
testFacStr <- encodeFactor(testFac)
testFacOut <- decodeFactor(testFacStr)
testthat::expect_equivalent(testFac, testFacOut)

}
