% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{devROGER}
\alias{devROGER}
\title{Get connection to devel database}
\usage{
devROGER(dest = "/pstore/data/bi/apps/ROGER/protROGER-v0.2.db")
}
\arguments{
\item{dest}{SQLite3 file which holds the ROGER database. The default value is the prototype database.}
}
\description{
Get connection to devel database
}
\examples{
devROGER()
}
