% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{dfLogical2Int}
\alias{dfLogical2Int}
\title{Convert logical values in a data.frame into integers}
\usage{
dfLogical2Int(df)
}
\arguments{
\item{df}{A data.frame}
}
\description{
Convert logical values in a data.frame into integers
}
\examples{
exampleDf <- data.frame(Teams=c("HSV", "FCB", "FCB", "HSV"),
             Player=c("Mueller", "Mueller", "Robben", "Holtby"),
             IsPlaying=c(TRUE, FALSE, TRUE, TRUE),
             Scores=c(3.5, 1.5, 1.5, 1.0), stringsAsFactors=TRUE)
intDf <- dfLogical2Int(exampleDf)
stopifnot(identical(intDf[,3], c(1L, 0L, 1L, 1L)))

}

