% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{importEdgeResult}
\alias{importEdgeResult}
\title{Import the content of a EdgeResult object into the database}
\usage{
importEdgeResult(conn, edgeResult, edgeObject, phenoData = NULL, xref = NA,
  anno = FALSE, enrichTbl, verbose = TRUE)
}
\arguments{
\item{conn}{Database connection}

\item{edgeResult}{An object of the class \code{\link[ribiosNGS]{EdgeResult-class}}}

\item{edgeObject}{An object of EdgeObject}

\item{phenoData}{An (annotated) data.frame containing sample annotations}

\item{xref}{External reference}

\item{anno}{Logical, whether annotation was provided by the edgeR script}

\item{enrichTbl}{Gene-set enrichment table}

\item{verbose}{Logical, whether verbose or not

The script is used by ngsDge_edgeR.Rscript to import edgeResult into ROGeR. It is very
experimental: the code has not been thoroughly tested and needs refactoring. Use with caution.}
}
\description{
Import the content of a EdgeResult object into the database
}
