% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{importLimmaResult}
\alias{importLimmaResult}
\title{Import the content of a EdgeResult object into the database}
\usage{
importLimmaResult(conn, eSet, eSetFit, enrichTbl, valType = c("RMA", "MAS5"),
  xref = NA, anno = FALSE, verbose = TRUE)
}
\arguments{
\item{conn}{Database connection}

\item{eSet}{An object of the class \code{\link[Biobase]{eSet-class}}}

\item{eSetFit}{An object of the class \code{\link[limma]{MArrayLM-class}}}

\item{enrichTbl}{Gene-set enrichment table}

\item{valType}{Value type}

\item{xref}{External reference}

\item{anno}{Logical, whether annotation was provided by the limma script}

\item{verbose}{Logical, whether verbose or not

The script is used by maDge_limma.Rscript to import limma results into ROGeR. It is very
experimental: the code has not been thoroughly tested and needs refactoring. Use it with caution.}
}
\description{
Import the content of a EdgeResult object into the database
}
