% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{limmaDgeTable}
\alias{limmaDgeTable}
\title{limma DGE table}
\usage{
limmaDgeTable(eFit, contrast = NULL)
}
\arguments{
\item{eFit}{An \code{MArrayLM} object}

\item{contrast}{Contrast(s) of choice}
}
\description{
limma DGE table
}
