% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{limmaDgeTables}
\alias{limmaDgeTables}
\title{limma DGE table}
\usage{
limmaDgeTables(eFit)
}
\arguments{
\item{eFit}{An \code{MArrayLM} object}

\item{contrast}{Contrast(s) of choice}
}
\description{
limma DGE table
}
