% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{limmaOneDgeTable}
\alias{limmaOneDgeTable}
\title{limma DGE table of one contrast}
\usage{
limmaOneDgeTable(eFit, contrast)
}
\arguments{
\item{eFit}{An \code{MArrayLM} object}

\item{contrast}{Contrast of choice}
}
\description{
limma DGE table of one contrast
}
