% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{maPhenoData}
\alias{maPhenoData}
\title{Return pheno data required by ROGER database}
\usage{
maPhenoData(eset, phenoDataFile = NULL, groups = NULL)
}
\arguments{
\item{eset}{An \code{ExpressionSet} object}

\item{phenoDataFile}{A pheno data file}

\item{groups}{Groups of the samples, if available}
}
\value{
The \code{ExpressionSet} object, with phenoData modified
}
\description{
Return pheno data required by ROGER database
}
