% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{maxDatasetID}
\alias{maxDatasetID}
\title{Max and new dataset ID}
\usage{
maxDatasetID(conn)
}
\arguments{
\item{conn}{Database connection}
}
\value{
Integer, max dataset ID
}
\description{
Max and new dataset ID
}
