% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{maxDesignID}
\alias{maxDesignID}
\title{Max and new design ID}
\usage{
maxDesignID(conn)
}
\arguments{
\item{conn}{Database connection}
}
\value{
Max design ID and new design ID (max+1)
}
\description{
Max and new design ID
}

