% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{newContrastIDs}
\alias{newContrastIDs}
\title{Return new contrast IDs}
\usage{
newContrastIDs(conn, contrastMatrix)
}
\arguments{
\item{conn}{Database connection}

\item{contrastMatrix}{A contrast matrix, its ncol will be used to extend the IDs}
}
\description{
Return new contrast IDs
}
