% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{newDatasetDesignID}
\alias{newDatasetDesignID}
\title{New design ID given an existing datasetID}
\usage{
newDatasetDesignID(conn, datasetID)
}
\arguments{
\item{conn}{Database connection}

\item{datasetID}{An existing datasetID}
}
\value{
New design ID (max+1)
}
\description{
New design ID given an existing datasetID
}
