% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{serializeMatrixByCol}
\alias{serializeMatrixByCol}
\title{Serialize a matrix by columns}
\usage{
serializeMatrixByCol(matrix)
}
\arguments{
\item{matrix}{A matrix}
}
\value{
A list of serializsed column values
}
\description{
Serialize a matrix by columns
}
\examples{
serializeMatrixByCol(matrix(1:4, ncol=2))
}

