% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{maxDatasetID}
\alias{maxDatasetID}
\title{Max and new dataset ID}
\usage{
maxDatasetID(conn)
}
\arguments{
\item{conn}{Database connection}
}
\value{
Max dataset ID and new dataset ID (max+1)
}
\description{
Max and new dataset ID
}

